refactor: reorganize type files
