+++ packages/vx-shape/test/Bar.test.tsx
@@ -19,6 +19,7 @@ describe('<Bar />', () => {
   });
 
   test('it should expose its ref via an innerRef prop', () => {
+    // eslint-disable-next-line jest/no-test-return-statement
     return new Promise(done => {
       const refCallback = (ref: SVGRectElement) => {
         expect(ref.tagName).toMatch('rect');
