+++ packages/vx-shape/test/BarGroupHorizontal.test.tsx
@@ -31,23 +31,6 @@ const y0 = () => 5;
 const y0Scale = scaleBand({ domain: [0, 100], range: [0, 100] });
 const y1Scale = scaleBand({ domain: [0, 100], range: [0, 100] });
 const xScale = scaleLinear({ domain: [0, 100], range: [0, 100] });
-
-// const y0 = () => 1;
-// const y0Scale = () => 2;
-// y0Scale.bandwidth = () => 10;
-// y0Scale.domain = () => [0, 100] as [number, number];
-// y0Scale.range = () => [0, 100] as [number, number];
-// y0Scale.copy = () => y0Scale;
-// const y1Scale = () => 1;
-// y1Scale.bandwidth = () => 2;
-// y1Scale.domain = () => [0, 100] as [number, number];
-// y1Scale.range = () => [0, 100] as [number, number];
-// y1Scale.copy = () => y1Scale;
-// const xScale = (d: Datum) => 5;
-// xScale.domain = () => [0, 100] as [number, number];
-// xScale.range = () => [0, 100] as [number, number];
-// xScale.copy = () => xScale;
-
 const color = () => 'violet';
 const keys = ['New York', 'San Francisco', 'Austin'];
 const width = 1;
