+++ packages/vx-shape/test/BarRounded.test.tsx
@@ -20,6 +20,7 @@ describe('<BarRounded />', () => {
   });
 
   test('it should expose its ref via an innerRef prop', () => {
+    // eslint-disable-next-line jest/no-test-return-statement
     return new Promise(done => {
       const refCallback = (ref: SVGPathElement) => {
         expect(ref.tagName).toMatch('path');
